feat: reorganize fish configuration scripts for better structure
- Moved abbreviation definitions under an interactive check in abbr.fish
- Nested zoxide initialization within an interactive check in zoxide.fish
- Removed  fish-abbreviation-tips plugin from fish_plugins (too distracting)

+++ home/dot_config/private_fish/conf.d/abbr.fish
@@ -1,18 +1,20 @@
-abbr -a -- sysu 'systemctl --user'
-abbr -a -- cls clear
-abbr -a cm chezmoi
+if status --is-interactive
+    abbr -a -- sysu 'systemctl --user'
+    abbr -a -- cls clear
+    abbr -a cm chezmoi
 
-abbr -a -- df duf
+    abbr -a -- df duf
 
-# git abbreviations
-abbr -a gco 'git checkout'
-abbr -a ga 'git add'
-abbr -a gcm 'git commit -m'
+    # git abbreviations
+    abbr -a gco 'git checkout'
+    abbr -a ga 'git add'
+    abbr -a gcm 'git commit -m'
 
-# package manager abbreviations
-abbr -a pi 'paru -S'
-abbr -a pu 'paru -Syu'
-abbr -a cleanup 'sudo pacman -Rns (pacman -Qtdq)'
+    # package manager abbreviations
+    abbr -a pi 'paru -S'
+    abbr -a pu 'paru -Syu'
+    abbr -a cleanup 'sudo pacman -Rns (pacman -Qtdq)'
 
-# journal abbreviations
-abbr -a jctl 'journalctl -p 3 -xb'
+    # journal abbreviations
+    abbr -a jctl 'journalctl -p 3 -xb'
+end
